pyproject.toml
[tool.poetry]
name = "podscript-mvp"
version = "0.1.0"
description = "MVP modular backend for podcast transcription"
authors = ["Manus AI <dev@manus.ai>"]
readme = "README.md"
packages = [
  { include = "podscript_api", from = "src" },
  { include = "podscript_pipeline", from = "src" },
  { include = "podscript_shared", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
pydantic = "^2.8.0"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
oss2 = "^2.18.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.10.0"
isort = "^5.13.0"
flake8 = "^7.1.0"
mypy = "^1.11.0"
httpx = "^0.27.0"


[tool.pytest.ini_options]
addopts = "--cov=src/podscript_api --cov=src/podscript_pipeline --cov=src/podscript_shared --cov-report=term-missing --cov-fail-under=80"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.coverage.run]
omit = [
  "src/podscript_pipeline/tingwu_adapter.py"
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover"
]